requirements.txt
wyze-sdk>=2.1.0
python-dotenv>=1.0.0
resend>=2.0.0
